backend/requirements.txt
fastapi
uvicorn[standard]
pyodbc
python-dotenv
azure-storage-blob
azure-identity
